+++ CHANGELOG.md
@@ -7,6 +7,8 @@
  - Added a command for .sp file generation from a template.
  - Added a command for README.md file generation from a template.
  - Added a command for github Actions file generation from a template.
+ - Added a command to generate a Project from a template.
+ - Added a linter based on spcomp.
  - Added more settings to support the above features.
 
 ### 1.7.1